requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
deepl==1.6.0
pip==22.0.4
requests==2.27.1
setuptools==60.2.0
urllib3==1.26.9
wheel==0.37.1
python-dotenv~=0.20.0
tqdm~=4.64.0